requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
Flask==0.12.2
Flask-Limiter==1.0.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
limits==1.3
MarkupSafe==1.0
psycopg2-binary==2.7.4
python-dotenv==0.8.2
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
SQLAlchemy-Searchable==1.0.3
SQLAlchemy-Utils==0.33.2
urllib3==1.22
validators==0.12.1
Werkzeug==0.14.1
WTForms==2.1
